prevent Class.getSimpleName() from throwing for ApacheHttpClient helper classes
