+++ dd-java-agent/instrumentation/apache-httpclient-4/src/main/java/datadog/trace/instrumentation/apachehttpclient/ApacheHttpClientInstrumentation.java
@@ -4,11 +4,7 @@ import static datadog.trace.agent.tooling.ClassLoaderMatcher.hasClassesNamed;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.DDElementMatchers.implementsInterface;
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
-import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.propagate;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
-import static datadog.trace.instrumentation.apachehttpclient.ApacheHttpClientDecorator.DECORATE;
-import static datadog.trace.instrumentation.apachehttpclient.ApacheHttpClientDecorator.HTTP_REQUEST;
-import static datadog.trace.instrumentation.apachehttpclient.HttpHeadersInjectAdapter.SETTER;
 import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.not;
@@ -19,8 +15,7 @@ import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.bootstrap.CallDepthThreadLocalMap;
 import datadog.trace.bootstrap.instrumentation.api.AgentScope;
-import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
-import java.io.IOException;
+
 import java.util.HashMap;
 import java.util.Map;
 import net.bytebuddy.asm.Advice;
@@ -30,8 +25,6 @@ import net.bytebuddy.implementation.bytecode.assign.Assigner;
 import net.bytebuddy.matcher.ElementMatcher;
 import org.apache.http.HttpHost;
 import org.apache.http.HttpRequest;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.ResponseHandler;
 import org.apache.http.client.methods.HttpUriRequest;
@@ -60,8 +53,8 @@ public class ApacheHttpClientInstrumentation extends Instrumenter.Tracing {
       packageName + ".ApacheHttpClientDecorator",
       packageName + ".HttpHeadersInjectAdapter",
       packageName + ".HostAndRequestAsHttpUriRequest",
-      getClass().getName() + "$HelperMethods",
-      getClass().getName() + "$WrappingStatusSettingResponseHandler",
+      packageName + ".HelperMethods",
+      packageName + ".WrappingStatusSettingResponseHandler",
     };
   }
 
@@ -150,44 +143,6 @@ public class ApacheHttpClientInstrumentation extends Instrumenter.Tracing {
         ApacheHttpClientInstrumentation.class.getName() + "$RequestWithHandlerAdvice");
 
     return transformers;
-  }
-
-  public static class HelperMethods {
-    public static AgentScope doMethodEnter(final HttpUriRequest request) {
-      final AgentSpan span = startSpan(HTTP_REQUEST);
-      span.setMeasured(true);
-      final AgentScope scope = activateSpan(span);
-
-      DECORATE.afterStart(span);
-      DECORATE.onRequest(span, request);
-
-      final boolean awsClientCall = request.getHeaders("amz-sdk-invocation-id").length > 0;
-      // AWS calls are often signed, so we can't add headers without breaking the signature.
-      if (!awsClientCall) {
-        propagate().inject(span, request, SETTER);
-      }
-      return scope;
-    }
-
-    public static void doMethodExit(
-        final AgentScope scope, final Object result, final Throwable throwable) {
-      if (scope == null) {
-        return;
-      }
-      final AgentSpan span = scope.span();
-      try {
-        if (result instanceof HttpResponse) {
-          DECORATE.onResponse(span, (HttpResponse) result);
-        } // else they probably provided a ResponseHandler.
-
-        DECORATE.onError(span, throwable);
-        DECORATE.beforeFinish(span);
-      } finally {
-        scope.close();
-        span.finish();
-        CallDepthThreadLocalMap.reset(HttpClient.class);
-      }
-    }
   }
 
   public static class UriRequestAdvice {
@@ -314,23 +269,4 @@ public class ApacheHttpClientInstrumentation extends Instrumenter.Tracing {
     }
   }
 
-  public static class WrappingStatusSettingResponseHandler implements ResponseHandler {
-    final AgentSpan span;
-    final ResponseHandler handler;
-
-    public WrappingStatusSettingResponseHandler(
-        final AgentSpan span, final ResponseHandler handler) {
-      this.span = span;
-      this.handler = handler;
-    }
-
-    @Override
-    public Object handleResponse(final HttpResponse response)
-        throws ClientProtocolException, IOException {
-      if (null != span) {
-        DECORATE.onResponse(span, response);
-      }
-      return handler.handleResponse(response);
-    }
-  }
 }
